requirements.txt
elasticsearch==8.11.1
tqdm==4.66.1
python-dotenv==1.0.0
python-telegram-bot==20.7
SQLAlchemy==2.0.25
aiosqlite==0.19.0
aiohttp==3.9.1 